202202252034 - Commit
(Monster.cpp) Implicit conversion was switched to explicit conversion in the Monster conversion constructor for the ACVariance of the dragon Monster object. Implicit conversion was throwing errors for those on Visual Studio.

(Item.h/cpp) Added a default value for the classification member variable in the Item class header file

(main) Added player.levelUp to the main function

(Player.cpp) Refactored the Player conversion constructor with initialized default values

(Object.cpp) Refactored the Object conversion constructor with initialized default values

+++ Monster.cpp
@@ -22,7 +22,7 @@ Monster::Monster( const Object& player ): Object()
         case Object::Type::slime:
             strengthVariance  = level  * 1.5;
             healthVariance  = level  * 1.25;
-            ACVariance = level / 2;
+            ACVariance = level / 2.0;
             break;
         case Object::Type::orc:
             strengthVariance  = level  * 2.0;
@@ -37,7 +37,7 @@ Monster::Monster( const Object& player ): Object()
         case Object::Type::dragon:
             strengthVariance  = level  * 6.0;
             healthVariance  = (long long)level  * level  * 3.0;
-            ACVariance = level * 3;
+            ACVariance = (double)level * 3.0;
             break;
     }
     std::normal_distribution<double> randomStrength(strengthVariance, level  / 4.0);
